+++ client/bindings/python/CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0.0-rc.3 - 2023-MM-DD
 
+### Added
+
+- `Client::hash_transaction_essence()`;
+
 ### Changed
 
 - Changes from the Rust library;